+++ src/commands/SECTION/sectionCommandParser.ts
@@ -8,10 +8,12 @@ import { SectionTypes } from '../../types/SectionType';
 import type { string_markdown_text } from '../../types/typeAliases';
 import { keepUnused } from '../../utils/organization/keepUnused';
 import { knowledgeCommandParser } from '../KNOWLEDGE/knowledgeCommandParser';
-import type { $PipelineJson } from '../_common/types/CommandParser';
-import type { $TaskJson } from '../_common/types/CommandParser';
-import type { CommandParserInput } from '../_common/types/CommandParser';
-import type { PipelineTaskCommandParser } from '../_common/types/CommandParser';
+import type {
+    $PipelineJson,
+    $TaskJson,
+    CommandParserInput,
+    PipelineTaskCommandParser,
+} from '../_common/types/CommandParser';
 import type { SectionCommand } from './SectionCommand';
 
 /**
@@ -192,7 +194,7 @@ export const sectionCommandParser: PipelineTaskCommandParser<SectionCommand> = {
             knowledgeCommandParser.$applyToPipelineJson(
                 {
                     type: 'KNOWLEDGE',
-                    sourceContent: $taskJson.content, // <- TODO: [🐝][main] !!3 Work with KNOWLEDGE which not referring to the source file or website, but its content itself
+                    knowledgeSourceContent: $taskJson.content, // <- TODO: [🐝][main] !!3 Work with KNOWLEDGE which not referring to the source file or website, but its content itself
                 },
                 $pipelineJson,
             );