+++ src/scrapers/pdf/playground/pdf-scraper-playground.ts
@@ -46,7 +46,7 @@ async function playground() {
 
     const knowledge = await pdfScraper.scrape(
         await makeKnowledgeSourceHandler(
-            { sourceContent: example },
+            { knowledgeSourceContent: example },
             { fs: $provideFilesystemForNode() },
             { rootDirname },
         ),