+++ src/conversion/parsePipeline.ts
@@ -4,15 +4,15 @@ import type { ParameterCommand } from '../commands/PARAMETER/ParameterCommand';
 import { sectionCommandParser } from '../commands/SECTION/sectionCommandParser';
 import { getParserForCommand } from '../commands/_common/getParserForCommand';
 import { parseCommand } from '../commands/_common/parseCommand';
-import type { $PipelineJson } from '../commands/_common/types/CommandParser';
-import type { $TaskJson } from '../commands/_common/types/CommandParser';
-import type { CommandBase } from '../commands/_common/types/CommandParser';
-import type { PipelineHeadCommandParser } from '../commands/_common/types/CommandParser';
-import type { PipelineTaskCommandParser } from '../commands/_common/types/CommandParser';
-import { DEFAULT_BOOK_TITLE } from '../config';
-import { DEFAULT_TASK_TITLE } from '../config';
-import { ORDER_OF_PIPELINE_JSON } from '../constants';
-import { RESERVED_PARAMETER_NAMES } from '../constants';
+import type {
+    $PipelineJson,
+    $TaskJson,
+    CommandBase,
+    PipelineHeadCommandParser,
+    PipelineTaskCommandParser,
+} from '../commands/_common/types/CommandParser';
+import { DEFAULT_BOOK_TITLE, DEFAULT_TASK_TITLE } from '../config';
+import { ORDER_OF_PIPELINE_JSON, RESERVED_PARAMETER_NAMES } from '../constants';
 import { ParseError } from '../errors/ParseError';
 import { UnexpectedError } from '../errors/UnexpectedError';
 import { HIGH_LEVEL_ABSTRACTIONS } from '../high-level-abstractions/index';
@@ -24,9 +24,8 @@ import type { PipelineString } from '../pipeline/PipelineString';
 import { validatePipelineString } from '../pipeline/validatePipelineString';
 import type { ScriptLanguage } from '../types/ScriptLanguage';
 import { SUPPORTED_SCRIPT_LANGUAGES } from '../types/ScriptLanguage';
-import type { number_integer } from '../types/typeAliases';
-import type { number_positive } from '../types/typeAliases';
-import type { string_name } from '../types/typeAliases';
+import type { number_integer, number_positive, string_name } from '../types/typeAliases';
+import { deflatePipeline } from '../utils/editable/edit-pipeline-string/deflatePipeline';
 import { extractAllListItemsFromMarkdown } from '../utils/markdown/extractAllListItemsFromMarkdown';
 import { extractOneBlockFromMarkdown } from '../utils/markdown/extractOneBlockFromMarkdown';
 import { flattenMarkdown } from '../utils/markdown/flattenMarkdown';
@@ -118,48 +117,12 @@ export function parsePipeline(pipelineString: PipelineString): PipelineJson {
     }
     pipelineString = removeMarkdownComments(pipelineString);
     pipelineString = spaceTrim(pipelineString) as PipelineString;
+
     // <- TODO: [😧] `spaceTrim` should preserve discriminated type *(or at lease `PipelineString`)*
+    pipelineString = deflatePipeline(pipelineString);
 
     // ==============
-    // Note: 1️⃣◽2️⃣ Process flat pipeline
-
-    const isMarkdownBeginningWithHeadline = pipelineString.startsWith('# ');
-    const isLastLineReturnStatement = pipelineString.split('\n').pop()!.split('`').join('').startsWith('->');
-    // TODO: Also (double)check
-    // > const usedCommands
-    // > const isBlocksUsed
-    // > const returnStatementCount
-
-    const isFlatPipeline = !isMarkdownBeginningWithHeadline && isLastLineReturnStatement;
-
-    // console.log({ isMarkdownBeginningWithHeadline, isLastLineReturnStatement, isFlatPipeline });
-
-    if (isFlatPipeline) {
-        const pipelineStringLines = pipelineString.split('\n');
-        const returnStatement = pipelineStringLines.pop()!;
-        const prompt = spaceTrim(pipelineStringLines.join('\n'));
-        pipelineString = validatePipelineString(
-            spaceTrim(
-                (block) => `
-                    # ${DEFAULT_BOOK_TITLE}
-
-                    ## Prompt
-
-                    \`\`\`
-                    ${block(prompt)}
-                    \`\`\`
-
-                    ${returnStatement}
-                `,
-            ),
-        );
-        // <- TODO: Maybe use book` notation
-
-        // console.log(pipelineString);
-    }
-
-    // ==============
-    // Note: 1️⃣◽3️⃣ Parse the markdown
+    // Note: 1️⃣◽2️⃣ Parse the markdown
     pipelineString = flattenMarkdown(pipelineString) /* <- Note: [🥞] */;
     pipelineString = pipelineString.replaceAll(
         /`\{(?<parameterName>[a-z0-9_]+)\}`/gi,